backend/requirements.txt
requests~=2.22.0
Flask~=1.1.2
python-dotenv~=0.14.0